pyproject.toml
[project]
name = "auto-notes"
version = "0.1.0"
description = "SJTU Canvas 课程自动笔记助手：下载、转录与 AI 笔记生成"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask",
    "flask-cors",
    "requests",
    "python-dotenv",
]

[project.optional-dependencies]
ai-google = [
    "google-generativeai",
    "google-genai",
]
asr-faster-whisper = [
    "faster-whisper",
]
asr-funasr = [
    "funasr",
    "modelscope",
    "torch",
    "torchaudio",
]
asr-sherpa-onnx = [
    "numpy",
    "sherpa-onnx",
    "sherpa-onnx-bin",
]
browser-playwright = [
    "playwright",
]
